Changed Model.updateMatch()
updateMatch() now takes 5 parameters so that the current match ID can be
kept constant and updated internally. A call to
Model.advanceTournament() has also been added to the bottom of this
function. this is something that should always be done after a match has
been updated, therefore it seems unreasonable that it would be able to
be called from the menu and risk an error.