Beginning of Multi-Sport Functionality.
Adds the ability for the user to manage team and player information related
to soccer players.  Modifies the Model class to allow for either a wrestling
or soccer model to be created.  This allows for the reuse of some already
implemented code.

This code increases the capabilities of the application.  With further
development, more functionality will be given to the soccer management
portion.  This will open up the application to be used for multiple sports,
making it a more robust product.

Does not implement changes in UI yet, will work on for future project.

Resolves: gzdwsu#12